Add new BT GATT Transport & defmt support.
`uuid v1.11.0` added an additional `PartialEq<NonNullUuid>`
implementation for `Uuid` so the code in the BT GATT transport could
not determine the correct type for the call to `Into::into`. We fixed
that by explicitly fixing the type to `Uuid`.

Add defmt support to zenoh-buffers, zenoh-collections and zenoh-protocol crates

bt-gatt: Implement endpoint and unimplement listener